+++ School/Module/Learning tasks/LTCopySequence.cs
@@ -9,11 +9,11 @@ namespace GoodAI.Modules.School.LearningTasks
 {
     public class LTCopySequence : AbstractLearningTask<ManInWorld>
     {
-        private static readonly TSHintAttribute STOP_REQUEST = new TSHintAttribute("Stop request","",TypeCode.Decimal,0,1); //check needed;
+        private static readonly TSHintAttribute STOP_REQUEST = new TSHintAttribute("Stop request","",TypeCode.Single,0,1); //check needed;
 
         // True if teacher agent can spawn on different places
         // Teacher should not cover agent
-        private static readonly TSHintAttribute TEACHER_ON_DIFF_START_POSITION = new TSHintAttribute("Teacher on diff start position","",TypeCode.Decimal,0,1); //check needed;
+        private static readonly TSHintAttribute TEACHER_ON_DIFF_START_POSITION = new TSHintAttribute("Teacher on diff start position","",TypeCode.Single,0,1); //check needed;
 
         protected Random m_rndGen = new Random();
         protected int m_stepsSincePresented = 0;
